전자결재 1차 완료.
1. 사이드바 새결재 진행 시 작성 메뉴 이동
2. 휴가 신청 반려시 소모 휴가 복구
3. 휴가 신청서 작성시 로그인 유저의 잔여 휴가수 불러오기
4. 휴가신청서 상신시 사원 연차정보 테이블 사용연차, 잔여 연차 update

+++ mayo_project/src/main/java/com/group/mayo/employee/model/service/HolidayServiceImpl.java
@@ -10,12 +10,16 @@ import com.group.mayo.employee.domain.CommuteEmployee;
 import com.group.mayo.employee.domain.Holiday;
 import com.group.mayo.employee.domain.HolidayEmployee;
 import com.group.mayo.employee.model.dao.HolidayDao;
+import com.group.mayo.holidayEmp.dao.EmpHolidayDao;
 
 @Service
 public class HolidayServiceImpl implements HolidayService{
 	
 	@Autowired
 	private HolidayDao dao;
+	
+	@Autowired
+	private EmpHolidayDao ehdao;
 	
 //	페이지 띄우기
 	@Override
@@ -57,9 +61,20 @@ public class HolidayServiceImpl implements HolidayService{
 
 	@Override
 	public int deleteHoliday(String ea_no) {
-		return dao.deleteHoliday(ea_no);
-	}
-	
-	
+		Holiday holiday = dao.selectHoliday(ea_no);
+		double restoreCnt = holiday.getHd_count();
+		HolidayEmployee hemployee = new HolidayEmployee();
+		hemployee.setUsed_count(restoreCnt);//사용 연차 = 복구해야할 연차
+		hemployee.setEmp_no(holiday.getEmp_no()); //복구할 사원 번호
+		hemployee.setLeft_count(restoreCnt + hemployee.getLeft_count());// 잔여연차 = 사용연차 + 기존 잔여연차
+		
+		int restoreResult = ehdao.updateRestore(hemployee);
+		int result = 0;
+		if(restoreResult != 0) {
+			result=dao.deleteHoliday(ea_no);
+		}
+		
+		return result;
+	}	
 	
 }
